% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{check_te_species}
\alias{check_te_species}
\title{List all threatened or endangered species in a input data set, using the USFWS species list.}
\usage{
check_te_species(
  x,
  sciname_col,
  resolve_input_taxonomy = TRUE,
  listing_status = "all",
  domestic_only = TRUE,
  viewer_table = FALSE
)
}
\arguments{
\item{x}{(data frame) A data frame containing scientific names.}

\item{sciname_col}{(character or numeric) The name or index of the column containing scientific names (genus and specific epithet).}

\item{resolve_input_taxonomy}{(logical) Defaults to TRUE. TRUE means taxonomic names will be resolved to current names and record IDs using ITIS and GBIF. Set to FALSE if taxonomy has already been resolved using one or both of these databases.}

\item{listing_status}{(character) Defaults to "all". A user may choose from "listed" to view only listed species, "listed or proposed" to view listed species and species proposed for listing or "all" to view listed, proposed, petitioned, candidate, and species of concern.}

\item{domestic_only}{(logical) Defaults to TRUE, meaning that in loading the ECOS T&E species list, the function pulls only species that occur domestically. If set to FALSE, it returns all species in the ECOS list.}

\item{viewer_table}{(logical) Defaults to TRUE. Shows the returned table in the plot viewer in R, complete with links to the species profiles in the ECOS and ITIS/GBIF databases.}
}
\value{
The function returns a data frame with the name of each submitted record that matched a threatened or endangered species,
the matched scientific name, common name, taxonomic group, listing status under the Endangered Species Act, listing status category,
a description of where the species is considered threatened or endangered and its ECOS profile ID.
}
\description{
\code{check_te_species()} generates a data frame with threatened or endangered species in the input data set by matching scientific names to the USFWS's catalog of T&E species. First it pulls the USFWS's catalog of Threatened and Endangered species and resolves scientific names and record IDs using two taxonomic databases: Integrated Taxonomic Information System (ITIS) and Global Biodiversity Information Facility (GBIF). It also optionally resolves names in the input dataframe using ITIS and GBIF.
}
\details{
\code{check_te_species()} relies on an API call to the Environmental Conservation Online System (ECOS) at https://ecos.fws.gov/ecp/ and the Global Names Verifier (GNV) https://verifier.globalnames.org/ from Global Names Architecture.
}
\examples{
\dontrun{
check_te_species(x = sfcn_mammals,
                 species_col = "scientificName",
                 resolve_input_taxonomy = FALSE)
check_te_species(x = sfcn_mammals,
                 species_col = "scientificName",
                 listing_status = "listed",
                 domestic_only = FALSE)
}
}
