% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{load_te_species}
\alias{load_te_species}
\title{Load a full list of threatened and endangered species from the USFWS}
\usage{
load_te_species(
  force_refresh = FALSE,
  domestic_only = TRUE,
  return_object = FALSE
)
}
\arguments{
\item{force_refresh}{(logical) Defaults to FALSE. If set to TRUE, gets the current list of T&E species. Data is automatically refreshed if it is over one month old.}

\item{domestic_only}{(logical) Defaults to TRUE, meaning that the function loads data only for species that occur domestically. If set to FALSE, it returns all species.}

\item{return_object}{(logical) Defaults to FALSE.If set to TRUE, returns a data frame object of T&E species and resolved taxonomy.}
}
\value{
The function returns a data frame with all listed domestic (or foreign and domestic species, if requested by the user) in the ECOS database, and taxonomic names are record IDs resolved using ITIS and GBIF.
}
\description{
\code{load_te_species()} gets the USFWS's catalog of Threatened and Endangered species and
resolves scientific names and record IDs using two taxonomic databases: Integrated Taxonomic Information System (ITIS) and
Global Biodiversity Information Facility (GBIF). It then stores this data as a csv file in the user's cache. It also returns the dataframe.
}
\details{
\code{load_te_species()} relies on an API call to the Environmental Conservation Online System (ECOS) at https://ecos.fws.gov/ecp/ and
the Global Names Verifier (GNV) https://verifier.globalnames.org/ from Global Names Architecture. It writes the resulting dataframe as a csv
entitled "resolved_te_species.csv" to the user's cache. In order to ensure that the species list is current, the function checks if there is
already a file named "resolved_te_species.csv" in the user's cache, and if there is, it checks if it was last modified over 30 days ago. If the
existing data is over 30 days old, it overwrites it with new data. Data can be updated at any time using force_refresh.
}
