% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{check_te_species}
\alias{check_te_species}
\title{Get a list of threatened or endangered species in a data set}
\usage{
check_te_species(
  x,
  sciname_col,
  resolve_taxonomy = TRUE,
  listing_status = "all",
  viewer_table = FALSE
)
}
\arguments{
\item{x}{(data frame) A data frame containing scientific names.}

\item{sciname_col}{(character or numeric) The name or index of the column containing scientific names (genus and specific epithet).}

\item{resolve_taxonomy}{(logical) Defaults to TRUE. TRUE means taxonomic names will be resolved to current names and record IDs using ITIS and GBIF.
Set to FALSE if taxonomy has already been resolved using one or both of these databases.}

\item{listing_status}{(character) Defaults to "all". A user may choose from "listed" to view only listed species, "listed or proposed" to view listed species and species proposed for listing
or "all" to view listed, proposed, and species of concern.}

\item{viewer_table}{(logical) Defaults to FALSE. If set to TRUE, shows the returned table in the plot viewer in R, complete with links to the species profiles in the ECOS and ITIS/GBIF databases.}
}
\value{
The function returns a data frame with the name of each submitted record that matched a threatened or endangered species,
the matched scientific name, common name, taxonomic group, listing status under the Endangered Species Act, listing status category,
a description of where the species is considered threatened or endangered and its ECOS profile ID.
}
\description{
\code{check_te_species()} generates a data frame with threatened or endangered species in the input data set by matching scientific names to the USFWS's catalog of T&E species.
}
\details{
\code{check_te_species()} relies on a file cached in load_te_species(). If this file does not exist, the function will exit with an error.
}
